1. improve the delete api. 2. unifily ident of BE codes.

+++ server/middleware/lists.js
@@ -10,12 +10,12 @@ const db = require('../config/dbconnect');
 
 function getAllWordsByListId(req, res, next) {
   var id = req.params.id;
-  db.any("select l.list_id, l.list_title, lw.word_id, w.word, w.symbol_id, s.symbol_path, s.symbol_text "
-    + "from lists l inner join listwords lw on l.list_id=lw.list_id "
-    + "inner join words w on lw.word_id=w.word_id "
-    + "inner join symbols s on w.symbol_id=s.symbol_id "
-    + "where l.list_id=" + '\'' + id + '\'')
-    .then(function (data) {
+  db.any("select l.list_id, l.list_title, lw.word_id, w.word, w.symbol_id, s.symbol_path, s.symbol_text " +
+      "from lists l inner join listwords lw on l.list_id=lw.list_id " +
+      "inner join words w on lw.word_id=w.word_id " +
+      "inner join symbols s on w.symbol_id=s.symbol_id " +
+      "where l.list_id=" + '\'' + id + '\'')
+    .then(function(data) {
       if (data.length > 0) {
         res.status(200)
           .json(data);
@@ -26,58 +26,62 @@ function getAllWordsByListId(req, res, next) {
         console.log("ERROR (404)");
       }
     })
-    .catch(function (err) {
+    .catch(function(err) {
       return next(err);
     });
 }
 
 function getAllWordsByListName(req, res, next) {
   var lTitle = req.params.title;
-  db.any("select l.list_id, l.list_title, lw.word_id, w.word, w.symbol_id, s.symbol_path, s.symbol_text "
-         + "from lists l inner join listwords lw on l.list_id=lw.list_id "
-         + "inner join words w on lw.word_id=w.word_id "
-         + "inner join symbols s on w.symbol_id=s.symbol_id "
-         + "where l.list_title=" + '\'' + lTitle + '\'')
-    .then(function (data) {
+  db.any("select l.list_id, l.list_title, lw.word_id, w.word, w.symbol_id, s.symbol_path, s.symbol_text " +
+      "from lists l inner join listwords lw on l.list_id=lw.list_id " +
+      "inner join words w on lw.word_id=w.word_id " +
+      "inner join symbols s on w.symbol_id=s.symbol_id " +
+      "where l.list_title=" + '\'' + lTitle + '\'')
+    .then(function(data) {
       if (data.length > 0) {
         res.status(200)
           .json(data);
-          console.log("All words for list " + lTitle + " were sent.");
+        console.log("All words for list " + lTitle + " were sent.");
       } else {
         res.status(404)
-        .send("ERROR: List " + '\'' + lTitle + '\' ' + "not found");
+          .send("ERROR: List " + '\'' + lTitle + '\' ' + "not found");
         console.log("ERROR (404)");
       }
     })
-    .catch(function (err) {
+    .catch(function(err) {
       return next(err);
     });
 }
 
 function deleteWordByID(req, res, next) {
-  console.log("deleteWordByID starts");
   var targetListID = parseInt(req.params.list_id);
   var targetWordID = parseInt(req.params.word_id);
-  db.result('delete from listwords where list_id=$1 AND word_id=$2',[targetListID, targetWordID])
-  .then(function (result) {
-     res.status(200)
-       .json({
-         status: 'success',
-         message: `Removed ${result.rowCount} word`
-       });
-       console.log("Removed " + result.rowCount + " word");
-      //  res.status(404)
-      //  .send("ERROR: word_id: " + targetWordID + " not found in list_id: " + targetListID);
-      //  console.log("ERROR (404)");
-     })
-   .catch(function (err) {
-     return next(err);
-   });
+  var query = 'select * from listwords where list_id=' + targetListID + ' and word_id=' + targetWordID;
+  console.log(query);
+  db.any(query)
+    .then(function(data) {
+      if (data.length > 0) {
+        db.result('delete from listwords where list_id=$1 AND word_id=$2', [targetListID, targetWordID])
+          .then(function(result) {
+            res.status(200)
+              .json({
+                status: 'success',
+                message: `Removed ${result.rowCount} word`
+              });
+          })
+      } else {
+        res.status(404)
+          .send("The word does not exist in the list.");
+      }
+    })
+    .catch(function(err) {
+      return next(err);
+    });
 }
 
-
 module.exports = {
-    getAllWordsByListId: getAllWordsByListId,
-    getAllWordsByListName: getAllWordsByListName,
-    deleteWordByID: deleteWordByID
+  getAllWordsByListId: getAllWordsByListId,
+  getAllWordsByListName: getAllWordsByListName,
+  deleteWordByID: deleteWordByID
 }
